+++ pkg/google/google_opts.go
@@ -11,9 +11,9 @@ import (
 // GoogleOpts Google 搜索配置。
 type GoogleOpts struct {
 	Enabled      bool
-	Blocked      []string // 屏蔽域名
-	PerSec       int      // 每秒限流，默认 1
-	PerMin       int      // 每分钟限流，默认 10
+	Blocked      []string            // 屏蔽域名
+	PerSec       int                 // 每秒限流，默认 3
+	PerMin       int                 // 每分钟限流，默认 30
 	ProxyResolve proxy.ProxyResolver // 代理端点动态解析函数（每次请求实时获取）
 }